requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
mediapipe
numpy
pillow
starlette
httpx
python-dotenv